페이징 및 로그인 기능 추가완료, 권한부여 추가필요

+++ src/main/java/com/practiceprj1/controller/BoardController.java
@@ -3,6 +3,7 @@ package com.practiceprj1.controller;
 import com.practiceprj1.domain.Board;
 import com.practiceprj1.service.BoardService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -24,16 +25,18 @@ public class BoardController {
     }
 
     @PostMapping("add")
-    public String add(Board board) {
-        service.add(board);
+    public String add(Board board, Authentication authentication, RedirectAttributes rttr) {
+        service.add(board, authentication);
 
-        return "redirect:/board/add";
+        rttr.addAttribute("id", board.getId());
+        return "redirect:/board";
     }
 
     @GetMapping("")
-    public String home(@RequestParam(value = "page", defaultValue = "1")Integer page, Model model) {
+    public String home(@RequestParam(value = "page", defaultValue = "1") Integer page, Model model) {
         model.addAllAttributes(service.list(page));
-        return null;
+
+        return "board/home";
     }
 
     @GetMapping("board")
@@ -53,17 +56,20 @@ public class BoardController {
     }
 
     @PostMapping("modify")
-    public String modify(Board board, RedirectAttributes rttr) {
-        service.update(board);
+    public String modify(Board board, Authentication authentication, RedirectAttributes rttr) {
+        if (service.hasAccess(board.getId(), authentication)) {
+            service.update(board);
+        }
         rttr.addAttribute("id", board.getId());
 
         return "redirect:/board";
     }
 
     @PostMapping("delete")
-    public String delete(Integer id) {
-        service.delete(id);
-
+    public String delete(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.delete(id);
+        }
         return "redirect:/";
     }
 }
